requirements.txt
boto3 
python-dotenv 
openai 
reportlab 
